services/timetable-management-service/requirements.txt
fastapi
uvicorn
python-jose
python-dotenv
requests
